Plan editor: author-feedback pass - sketch-parity canvas, build wording, gap containment, half-cell markers
Canvas restyled to the sketch tool's design language: red major lines
removed, purple chunk-style dashed cell grid (--canvas-chunk), heavier
axis lines with a crosshair origin marker at (0,0), ruler-style purple
gap connectors/labels, theme-aware canvas text (legible in light mode),
stronger build-area fill. Zone is now Build in every user-facing string
(wire schema unchanged). Gap links require the crossing span to lie
inside the zone minus holes and cross no piece interior - a merely
abutting band or a span over land no longer links (isolated-spawn now
shows exactly its two 10-block bridges). Markers snap to the half-cell
lattice (marker = stamp centre, so 2.5-block offsets are valid
positions); plan 'at' offsets are doubles, integer JSON unchanged.
Pgm.Tests 224 and JS 103 green; Api plan-endpoint tests pass (the
shared-DB isolation failures reproduce identically on the baseline).

+++ src/PgmStudio.Client/Pages/Plan/PlanEditor.razor.cs
@@ -394,7 +394,7 @@ public partial class PlanEditor
         [JsonPropertyName("markerKind")] public string MarkerKind { get; set; } = "";
         [JsonPropertyName("index")] public int Index { get; set; }
         [JsonPropertyName("piece")] public string Piece { get; set; } = "";
-        [JsonPropertyName("at")] public int[]? At { get; set; }
+        [JsonPropertyName("at")] public double[]? At { get; set; }
         [JsonPropertyName("facing")] public string Facing { get; set; } = "";
     }
